Update for Sentiment Analysis
This contains updated ChatbotCode that reference custom model for sentiment analysis, as well as the code to create the optimal model between Decision Tree and Naive Bayes Classifier. A seed (random_state) is set for reproducibility due to pkl being large to upload. Testing and Training sets are also provided.